requirements.txt
Django
django-ninja
django-ninja-jwt[crypto]
django-cors-headers
pydantic[email]
rav
python-decouple